+++ src/FSM/SharedFSMs/E-Stop/EStopReleased.cpp
@@ -13,12 +13,12 @@ EStopReleased::~EStopReleased() {}
 
 void EStopReleased::entry(){
     PRINT_STATE;
-	data->sender->send_event(Topic::ACTUATOR, ActuatorEnum::TRAFFIC_RED_ON, EventPriority::DEFAULT);
-	data->sender->send_event(Topic::ACTUATOR, ActuatorEnum::TRAFFIC_YELLOW_OFF, EventPriority::DEFAULT);
+	data->sender->send_event((int8_t)Topic::ACTUATOR,(int8_t) ActuatorEnum::TRAFFIC_RED_ON, (int8_t)EventPriority::DEFAULT);
+	data->sender->send_event((int8_t)Topic::ACTUATOR,(int8_t) ActuatorEnum::TRAFFIC_YELLOW_OFF,(int8_t) EventPriority::DEFAULT);
 }
 
 void EStopReleased::exit(){
-	data->sender->send_event(Topic::ACTUATOR, ActuatorEnum::TRAFFIC_RED_OFF,EventPriority::DEFAULT);
+	data->sender->send_event((int8_t)Topic::ACTUATOR,(int8_t) ActuatorEnum::TRAFFIC_RED_OFF,(int8_t)EventPriority::DEFAULT);
     PRINT_STATE;
 }
